Tests/LibWeb/Text/expected/HTMLInputElement-edit-value.txt

input[type=text] value: PASS
input[type=search] value: PASS
input[type=tel] value: PASS
input[type=url] value: PASS
input[type=email] value: PASS
input[type=password] value: PASS
